% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statable.R
\name{stata_close_session}
\alias{stata_close_session}
\title{Close a Stata session}
\usage{
stata_close_session(session = stata_default_session())
}
\arguments{
\item{session}{The Stata session to close}
}
\description{
Close a Stata session
}
