% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stata_data_out}
\alias{stata_data_out}
\title{Transfer data from Stata to R}
\usage{
stata_data_out(..., session = stata_default_session())
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{session}{The Stata session to read the data set from}
}
\value{
A data.frame containing the current Stata data set
}
\description{
Transfer data from Stata to R
}
