% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statable.R
\name{stata_default_session}
\alias{stata_default_session}
\title{Get the default Stata session}
\usage{
stata_default_session()
}
\description{
Returns the default Stata session. It will be started if it is not already running.
}
