% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{stata_evaluate}
\alias{stata_evaluate}
\title{Evaluate Stata commands and return details of their evaluation}
\usage{
stata_evaluate(
  commands,
  ...,
  session = stata_default_session(),
  env = parent.frame()
)
}
\arguments{
\item{commands}{A vector of Stata commands to run}

\item{...}{These dots are for future extensions and must be empty.}

\item{session}{The Stata session to run the commands in}

\item{env}{The R environment in which to look for data frames to transfer to Stata}
}
\description{
\code{stata_evaluate} similar to the \code{\link[evaluate:evaluate]{evaluate::evaluate()}} function. It returns
an \code{evaluation} object that consists of the Stata commands run and the
corresponding output and errors.
}
