% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statable.R
\name{stata_path}
\alias{stata_path}
\title{Get or set the path to the Stata executable}
\usage{
stata_path(path = NULL)
}
\arguments{
\item{path}{The path to the Stata executable}
}
\description{
Get or set the path to the Stata executable
}
