% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statable.R
\name{stata_run}
\alias{stata_run}
\title{Run a Stata command}
\usage{
stata_run(commands, session = stata_default_session())
}
\arguments{
\item{commands}{A vector of Stata commands to run}

\item{session}{The Stata session to run the commands in}
}
\description{
Run a Stata command
}
